Adjust/remove tests, move `indices_to_mask` to `utils.__init__`. - Remove test_bagging.test_consistent_index_sampling - Update formatting checks in test_bagging.test_estimators_samples   to make sure each mask is numpy boolean array - Move indices_to_mask from utils.validation to utils.__init__
