+++ sklearn/ensemble/tests/test_bagging.py
@@ -665,22 +665,6 @@ def test_oob_score_removed_on_warm_start():
     assert_raises(AttributeError, getattr, clf, "oob_score_")
 
 
-def test_consistent_index_sampling():
-    # Make sure randomly drawn indices generated at fit time are identically 
-    # reproduced at a later time (for use in, e.g., OOB scoring)
-    X, y = make_hastie_10_2(n_samples=20, random_state=1)
-    
-    bagging = BaggingClassifier(KNeighborsClassifier(), max_samples=0.5,
-                                max_features=0.5)
-    bagging.fit(X, y)
-
-    feat_inds_list = bagging.estimators_features_
-    feat_samp_inds_gen = bagging._get_estimators_indices()
-
-    for feat_inds1, (feat_inds2, _) in zip(feat_inds_list, feat_samp_inds_gen):
-        assert_array_equal(feat_inds1, feat_inds2) 
-
-
 def test_oob_score_consistency():
     # Make sure OOB scores are identical when random_state, estimator, and 
     # training data are fixed and fitting is done twice
@@ -701,21 +685,22 @@ def test_estimators_samples():
                                 bootstrap=False)
     bagging.fit(X, y)
 
-    # Get attributes, convert possible generators to lists
-    estimators_samples = list(bagging.estimators_samples_)
+    # Get relevant attributes
+    estimators_samples = bagging.estimators_samples_
     estimators_features = bagging.estimators_features_
     estimators = bagging.estimators_
 
     # Test for correct formatting
     assert_equal(len(estimators_samples), len(estimators))
     assert_equal(len(estimators_samples[0]), len(X))
-    assert_true(isinstance(estimators_samples[0][0], (np.bool_, bool)))
+    assert_equal(estimators_samples[0].dtype.kind, 'b')
 
     # Re-fit single estimator to test for consistent sampling
     estimator_index = 0
     estimator_samples = estimators_samples[estimator_index]
     estimator_features = estimators_features[estimator_index]
     estimator = estimators[estimator_index]
+
     X_train = (X[estimator_samples])[:, estimator_features]
     y_train = y[estimator_samples]
